// src/schemas/account-api.ts
export interface components {
	schemas: {
// ...
		Service: {
			/** @example compute */
			name?: string;
			/**
			 * @description CRM: icdc_service.description
			 * @example Order cloud services directly from the catalog: virtual machines, application images, and network access.
			 */
			readonly description?: string;
			/** @example IBACloud Compute */
			display_name?: string;
			/** @example  */
			path?: string;
			/** @example https://compute.xby.icdc.io */
			url?: string;
			/** @example 3 */
			position?: number;
		};
// ...
	};
// ...
}